#536: Renamed `ReporterFactory` to `AbstractReporterFactory`.
