+++ api/src/test/java/org/itsallcode/openfasttrace/api/report/TestReporterFactory.java
@@ -26,7 +26,7 @@ class TestReporterFactory
         assertThat(reporterFactory.getContext(), sameInstance(context));
     }
 
-    private static class TestingReporterFactory extends ReporterFactory
+    private static class TestingReporterFactory extends AbstractReporterFactory
     {
         @Override
         public boolean supportsFormat(String format)
