+++ reporter/html/src/main/java/org/itsallcode/openfasttrace/report/html/HtmlReporterFactory.java
@@ -2,13 +2,14 @@ package org.itsallcode.openfasttrace.report.html;
 
 import org.itsallcode.openfasttrace.api.core.Trace;
 import org.itsallcode.openfasttrace.api.report.Reportable;
+import org.itsallcode.openfasttrace.api.report.AbstractReporterFactory;
 import org.itsallcode.openfasttrace.api.report.ReporterFactory;
 
 /**
  * A {@link ReporterFactory} for HTML reports. This supports the
  * <code>html</code> format.
  */
-public class HtmlReporterFactory extends ReporterFactory
+public class HtmlReporterFactory extends AbstractReporterFactory
 {
     private static final String HTML_REPORT_FORMAT = "html";
